refactoring into major structure before quitting for the night
