+++ tests/test_controller.py
@@ -4,7 +4,7 @@ import pytest
 import random
 
 from vent.common.message import SensorValues, ControlSetting
-from vent.alarms import AlarmSeverity, Alarm
+from vent.alarm import AlarmSeverity, Alarm
 from vent.common.values import ValueName
 from vent.coordinator.coordinator import get_coordinator
 from vent.controller.control_module import get_control_module
